Fix a concurrency issue with AutoSuspendLocker
The timer operations is used concurrently, but they are not
threadsafe. To fix this, we use a mutex for performing operations on
the timer.

Although this may decrease its performance, the AutoSuspendLocker is
likely not the bottleneck of any system. If it is, we can improve this
by using task.Idempotent or some other asynchronous message passing
trick to ensure approximately correct operation. This should be
sufficient in burst modes, as it's very likely that the timer would be
updated quickly afterwards.